HAPI-225 split up functions
